requirements.txt
pika==0.9.8
./xmpp.zip
sleekxmpp

